Issue #131
